Add swapchain, surface
- Define DIBS_DEBUG in Debug, use for trace().
- Add const to parameters in function definitions.

+++ include/dibs/vec2.hpp
@@ -7,7 +7,7 @@ struct tvec2 {
 	T x{}, y{};
 };
 
-using ivec2 = tvec2<std::uint32_t>;
-using uvec2 = tvec2<std::uint64_t>;
+using ivec2 = tvec2<std::int32_t>;
+using uvec2 = tvec2<std::uint32_t>;
 using vec2 = tvec2<float>;
 } // namespace dibs
